split three files under the 500-line architecture limit
Pure moves into partial-class siblings following each area's existing
pattern: the GDS session's draft building into .Drafts.cs, Component's
slider management into Component.Sliders.cs, and the import dialog's
result summary into .Summary.cs. No behavior change; unblocks the
hard-limit architecture test on this branch.
